+++ NEWS.md
@@ -1,4 +1,4 @@
-# ruler 0.1.4.9000
+# ruler 0.2.0
 
 This version is reaction to changes in `dplyr` 0.8.0.